+++ src/layouts/GridItem/GridItem.layout.tsx
@@ -1,9 +1,15 @@
+import { forwardRef } from "react";
 import styled from "styled-components";
 
 import { gridItemStyles } from "./GridItem.styles";
 import { useGridItemPosition } from "./GridItem.utils";
 
-import { useClassName, useStyleProps, useTestId } from "@utils";
+import {
+  useClassName,
+  useHtmlAttribute,
+  useStyleProps,
+  useTestId,
+} from "@utils";
 
 import type {
   GridItemDefaultProps,
@@ -25,12 +31,13 @@ const defaultGridItemProps: GridItemDefaultProps = {
   children: null,
   className: null,
   as: "div",
+  styleOverride: null,
 };
 
 /**
  * Renders a grid item with the specified position and children.
  *
- * @version 0.0.3
+ * @version 0.0.4
  *
  * @param {GridItemProps} props - The props for the grid item.
  * @param {Optional<GridStartPosition>} props.rowStart - The start position of the row.
@@ -42,31 +49,36 @@ const defaultGridItemProps: GridItemDefaultProps = {
  * @param {ReactNode} props.children - The children to render inside the grid item.
  * @return {JSX.Element} The rendered grid item.
  */
-const GridItem = (
-  props: GridItemProps & TestIdProps = defaultGridItemProps
-): JSX.Element => {
-  const testId = useTestId("grid-item", props as TestIdProps);
-  const position = useGridItemPosition(defaultGridItemProps, props);
-  const styleProps = useStyleProps(position);
-  const className = useClassName(props);
-
-  return (
-    <GridItemContainer
-      {...styleProps}
-      as={props.as ?? "div"}
-      data-testid={testId}
-      className={className}
-      data-layout="grid-item"
-      data-column-start={position.columnStart}
-      data-column-end={position.columnEnd}
-      data-row-start={position.rowStart}
-      data-row-end={position.rowEnd}
-      data-fill={position.fill}
-    >
-      {props.children}
-    </GridItemContainer>
-  );
-};
+const GridItem = forwardRef<HTMLDivElement, GridItemProps & TestIdProps>(
+  (
+    props: GridItemProps & TestIdProps = defaultGridItemProps,
+    ref
+  ): JSX.Element => {
+    const testId = useTestId("grid-item", props as TestIdProps);
+    const position = useGridItemPosition(defaultGridItemProps, props);
+    const styleProps = useStyleProps(position);
+    const className = useClassName(props);
+    const style = useHtmlAttribute(props.styleOverride);
+    return (
+      <GridItemContainer
+        {...styleProps}
+        as={props.as ?? "div"}
+        data-testid={testId}
+        className={className}
+        style={style}
+        ref={ref}
+        data-layout="grid-item"
+        data-column-start={position.columnStart}
+        data-column-end={position.columnEnd}
+        data-row-start={position.rowStart}
+        data-row-end={position.rowEnd}
+        data-fill={position.fill}
+      >
+        {props.children}
+      </GridItemContainer>
+    );
+  }
+);
 GridItem.defaultProps = defaultGridItemProps;
 
 export { GridItem };