+++ src/layouts/Table/Table.layout.tsx
@@ -1,3 +1,4 @@
+import { forwardRef } from "react";
 import styled from "styled-components";
 
 import { tableLayoutStyles } from "./Table.styles";
@@ -7,38 +8,54 @@ import {
   TableStyleProps,
 } from "./Table.types";
 
-import { useTestId, useMergedProps, useClassName, useStyleProps } from "@utils";
+import {
+  useTestId,
+  useMergedProps,
+  useClassName,
+  useStyleProps,
+  useHtmlAttribute,
+} from "@utils";
 
 import type { TestIdProps } from "@types";
 
 const defaultTableProps: TableDefaultProps = {
   children: null,
   className: null,
+  styleOverride: null,
   layout: "auto",
 };
 
 /**
  * A structured layout element used to display data in rows and columns.
  *
- * @version 0.0.2
+ * @version 0.0.3
  *
  * @param {TableProps & TestIdProps} props - Table component props
  * @returns {JSX.Element}
  */
-const Table = (props: TableProps & TestIdProps): JSX.Element => {
-  const mergedProps = useMergedProps(defaultTableProps, props);
-  const testId = useTestId("table", props);
+const Table = forwardRef<HTMLTableElement, TableProps & TestIdProps>(
+  (props: TableProps & TestIdProps, ref): JSX.Element => {
+    const mergedProps = useMergedProps(defaultTableProps, props);
+    const testId = useTestId("table", props);
 
-  const className = useClassName(mergedProps);
+    const className = useClassName(mergedProps);
 
-  const styleProps = useStyleProps(mergedProps);
+    const styleProps = useStyleProps(mergedProps);
+    const style = useHtmlAttribute(mergedProps.styleOverride);
 
-  return (
-    <TableLayout data-testid={testId} className={className} {...styleProps}>
-      {mergedProps.children}
-    </TableLayout>
-  );
-};
+    return (
+      <TableLayout
+        data-testid={testId}
+        className={className}
+        style={style}
+        ref={ref}
+        {...styleProps}
+      >
+        {mergedProps.children}
+      </TableLayout>
+    );
+  }
+);
 Table.defaultProps = defaultTableProps;
 Table.displayName = "Table";